requirements.txt
pygame
python-chess
pygame_textinput